requirements.txt
flask
twilio
openai
python-dotenv
